Choi

+++ team_mypet/src/main/java/com/jeonju/mypet/controller/CommunityController.java
@@ -143,13 +143,14 @@ public class CommunityController {
 		int replyCount = commuService.getReplyCount(cm_idx);
 		int plusCmView = commuService.plusCmView(cm_idx);
 		
-		long midx = 0;
+		int midx = 0;
+//		long midx = 0;
 		
 		HttpSession session = request.getSession();
 		
 		if(session.getAttribute("midx") != null ) {
-//			midx = (int)session.getAttribute("midx");
-			midx = (long) session.getAttribute("midx");
+			midx = (int)session.getAttribute("midx");
+//			midx = (long) session.getAttribute("midx");
 			
 			Community_likeVo community_likeVo = new Community_likeVo();
 			community_likeVo.setMidx((int) midx);
